Updated for NTR-HR2 changes

+++ ffmpeg_opt/libavcodec/put_bits.h
@@ -224,7 +224,7 @@ static inline void put_bits_no_assert(PutBitContext *s, int n, BitBuf value)
     } else {
         bit_buf   <<= bit_left;
         bit_buf    |= value >> (n - bit_left);
-        if (s->buf_end - s->buf_ptr >= sizeof(BitBuf)) {
+        if (s->buf_end - s->buf_ptr >= (intptr_t)sizeof(BitBuf)) {
             AV_WBBUF(s->buf_ptr, bit_buf);
             s->buf_ptr += sizeof(BitBuf);
         } else {
@@ -262,7 +262,7 @@ static inline void put_bits_le(PutBitContext *s, int n, BitBuf value)
 
     bit_buf |= value << (BUF_BITS - bit_left);
     if (n >= bit_left) {
-        if (s->buf_end - s->buf_ptr >= sizeof(BitBuf)) {
+        if (s->buf_end - s->buf_ptr >= (intptr_t)sizeof(BitBuf)) {
             AV_WLBUF(s->buf_ptr, bit_buf);
             s->buf_ptr += sizeof(BitBuf);
         } else {
@@ -314,7 +314,7 @@ static void av_unused put_bits32(PutBitContext *s, uint32_t value)
 #else
     bit_buf     = (uint64_t)bit_buf << bit_left;
     bit_buf    |= (BitBuf)value >> (BUF_BITS - bit_left);
-    if (s->buf_end - s->buf_ptr >= sizeof(BitBuf)) {
+    if (s->buf_end - s->buf_ptr >= (intptr_t)sizeof(BitBuf)) {
         AV_WBBUF(s->buf_ptr, bit_buf);
         s->buf_ptr += sizeof(BitBuf);
     } else {
